fix room

+++ app/src/main/java/com/binar/sciroper/data/db/user/UserDAO.kt
@@ -10,10 +10,10 @@ interface UserDAO {
     suspend fun insertUser(user: User): Long
 
     @Delete
-    suspend fun deleteUser(user: User): Long
+    suspend fun deleteUser(user: User): Int
 
     @Update
-    suspend fun updateUser(user: User): Long
+    suspend fun updateUser(user: User): Int
 
     @Query("SELECT * FROM ${User.TABLE_NAME} WHERE id = :id")
     fun getUserById(id: Int): LiveData<User>
@@ -22,7 +22,7 @@ interface UserDAO {
     fun getUserByEmailAndPassword(email: String, password: String): LiveData<User>
 
     @Query("DELETE FROM ${User.TABLE_NAME}")
-    suspend fun clearAllUser(): Long
+    suspend fun clearAllUser(): Int
 
     @Query("SELECT * FROM ${User.TABLE_NAME}")
     fun getUsers(): LiveData<List<User>>
